move shieldtrc20 case to http dir and change gradle test thread to 2
