+++ testcase/build.gradle
@@ -174,7 +174,7 @@ task dailyBuild(type: Test) {
     useTestNG {
         suites(file('src/test/resources/daily-build.xml'))
         parallel 'tests'
-        threadCount 1
+        threadCount 2
     }
 }
