+++ testcase/src/test/java/stest/tron/wallet/dailybuild/http/HttpShieldTrc20Token002.java
@@ -1,4 +1,4 @@
-package stest.tron.wallet.dailybuild.zentrc20token;
+package stest.tron.wallet.dailybuild.http;
 
 import com.alibaba.fastjson.JSONArray;
 import com.alibaba.fastjson.JSONObject;
