+++ testcase/src/test/java/stest/tron/wallet/dailybuild/http/HttpShieldTrc20Token003.java
@@ -1,4 +1,4 @@
-package stest.tron.wallet.dailybuild.zentrc20token;
+package stest.tron.wallet.dailybuild.http;
 
 import com.alibaba.fastjson.JSONArray;
 import com.alibaba.fastjson.JSONObject;
@@ -13,7 +13,7 @@ import stest.tron.wallet.common.client.utils.HttpMethed;
 import stest.tron.wallet.common.client.utils.ZenTrc20Base;
 
 @Slf4j
-public class HttpShieldTrc20Token004 extends ZenTrc20Base {
+public class HttpShieldTrc20Token003 extends ZenTrc20Base {
 
   JSONArray shieldedReceives = new JSONArray();
   String txid;
@@ -24,8 +24,6 @@ public class HttpShieldTrc20Token004 extends ZenTrc20Base {
       .getStringList("httpnode.ip.list").get(1);
   private String httpSolidityNode = Configuration.getByPath("testng.conf")
       .getStringList("httpnode.ip.list").get(2);
-  private String httpPbftNode = Configuration.getByPath("testng.conf")
-          .getStringList("httpnode.ip.list").get(4);
   private JSONObject responseContent;
   private HttpResponse response;
   private JSONObject shieldAccountInfo1;
@@ -38,13 +36,13 @@ public class HttpShieldTrc20Token004 extends ZenTrc20Base {
   private Long account1Receive1V2Amount = 10L;
   private Long account2Receive1V2Amount = publicFromAmount - account1Receive1V2Amount;
   private Long account1Receive2V2Amount = 13L;
-  private Long account2Receive2V2Amount = publicFromAmount + account2Receive1V2Amount
-      - account1Receive2V2Amount;
+  private Long account2Receive2V2Amount = publicFromAmount
+      + account2Receive1V2Amount - account1Receive2V2Amount;
 
   /**
    * constructor.
    */
-  @BeforeClass(enabled = true, description = "Prepare for transfer without ask")
+  @BeforeClass(enabled = true, description = "Prepare for transfer")
   public void prepareForTransfer() {
     //Create two shield account
     response = getNewShieldedAddress(httpnode);
@@ -58,7 +56,6 @@ public class HttpShieldTrc20Token004 extends ZenTrc20Base {
     shieldedReceives.clear();
     shieldedReceives = getHttpShieldedReceivesJsonArray(shieldedReceives, publicFromAmount,
         shieldAccountInfo1.getString("payment_address"), getRcm(httpnode));
-    HttpMethed.waitToProduceOneBlock(httpnode);
     response = createShieldContractParameters(httpnode, publicFromAmount, shieldAccountInfo1,
         shieldedReceives);
     responseContent = HttpMethed.parseResponseContent(response);
@@ -70,7 +67,7 @@ public class HttpShieldTrc20Token004 extends ZenTrc20Base {
         zenTrc20TokenOwnerKey);
     HttpMethed.waitToProduceOneBlock(httpnode);
     HttpMethed.waitToProduceOneBlock(httpnode);
-    HttpMethed.waitToProduceOneBlock(httpnode);
+
     shieldedReceives.clear();
     shieldedReceives = getHttpShieldedReceivesJsonArray(shieldedReceives, publicFromAmount,
         shieldAccountInfo1.getString("payment_address"), getRcm(httpnode));
@@ -83,42 +80,31 @@ public class HttpShieldTrc20Token004 extends ZenTrc20Base {
         zenTrc20TokenOwnerAddressString, shieldAddress, mint, responseContent
             .getString("trigger_contract_input"), maxFeeLimit, 0L, 0, 0L,
         zenTrc20TokenOwnerKey);
+    HttpMethed.waitToProduceOneBlock(httpnode);
+    HttpMethed.waitToProduceOneBlock(httpnode);
 
-    HttpMethed.waitToProduceOneBlock(httpnode);
-    HttpMethed.waitToProduceOneBlock(httpnode);
   }
 
 
-  @Test(enabled = true, description = "Transfer type with 1V1 without ask by http")
-  public void test01TransferTypeWith1V1WithoutAskByHttp() {
+  @Test(enabled = true, description = "Transfer type with 1V1 by http")
+  public void test01TransferTypeWith1V1ByHttp() {
     account1IvkNoteTxs = scanShieldTrc20NoteByIvk(httpnode, shieldAccountInfo1);
     shieldSpends.clear();
     shieldSpends = createAndSetShieldedSpends(httpnode, shieldSpends, account1IvkNoteTxs
         .getJSONObject(0));
-    HttpMethed.waitToProduceOneBlock(httpnode);
     shieldedReceives.clear();
     shieldedReceives = getHttpShieldedReceivesJsonArray(shieldedReceives, publicFromAmount,
         shieldAccountInfo2.getString("payment_address"), getRcm(httpnode));
-    response = createShieldContractParametersWithoutAskForTransfer(httpnode, shieldAccountInfo1,
+    response = createShieldContractParametersForTransfer(httpnode, shieldAccountInfo1,
         shieldSpends, shieldedReceives);
-    JSONObject shieldedTrc20Parameters = HttpMethed.parseResponseContent(response);
-    HttpMethed.printJsonContent(shieldedTrc20Parameters);
-    JSONObject spendAuthSig = createSpendAuthSig(httpnode, shieldAccountInfo1,
-        shieldedTrc20Parameters.getString("message_hash"), account1IvkNoteTxs
-            .getJSONObject(0).getJSONObject("note").getString("rcm"));
-    HttpMethed.printJsonContent(spendAuthSig);
-    JSONArray spendAuthSigArray = new JSONArray();
-    spendAuthSigArray.add(spendAuthSig);
-
-    response = getTriggerInputForShieldedTrc20Contract(httpnode, shieldedTrc20Parameters,
-        spendAuthSigArray);
     responseContent = HttpMethed.parseResponseContent(response);
     HttpMethed.printJsonContent(responseContent);
 
     txid = HttpMethed.triggerContractGetTxidWithVisibleTrue(httpnode,anotherHttpnode,
         zenTrc20TokenOwnerAddressString, shieldAddress, transfer, responseContent
-            .getString("value"), maxFeeLimit, 0L, 0, 0L,
+            .getString("trigger_contract_input"), maxFeeLimit, 0L, 0, 0L,
         zenTrc20TokenOwnerKey);
+    HttpMethed.waitToProduceOneBlock(httpnode);
     HttpMethed.waitToProduceOneBlock(httpnode);
     HttpMethed.waitToProduceOneBlock(httpnode);
     response = HttpMethed.getTransactionInfoById(httpnode, txid, true);
@@ -127,16 +113,16 @@ public class HttpShieldTrc20Token004 extends ZenTrc20Base {
     Assert.assertTrue(responseContent.getJSONObject("receipt")
         .getLong("energy_usage_total") > 300000L);
     Assert.assertEquals(responseContent.getString("contract_address"), shieldAddress);
-    Assert.assertEquals(responseContent.getJSONObject("receipt").getString("result"), "SUCCESS");
+    Assert.assertEquals(responseContent.getJSONObject("receipt").getString("result"),
+        "SUCCESS");
 
     account1OvkNoteTxs = scanShieldTrc20NoteByOvk(httpnode, shieldAccountInfo1);
     logger.info(account1OvkNoteTxs.toJSONString());
     Assert.assertEquals(account1OvkNoteTxs.size(), 1);
   }
 
-  @Test(enabled = true, description = "Transfer type with 1V2 without ask by http")
-  public void test02TransferTypeWith1V2WithoutAskByHttp() {
-
+  @Test(enabled = true, description = "Transfer type with 1V2 by http")
+  public void test02TransferTypeWith1V2ByHttp() {
     account1IvkNoteTxs = scanShieldTrc20NoteByIvk(httpnode, shieldAccountInfo1);
 
     Assert.assertTrue(isShieldedTrc20ContractNoteSpent(httpnode, shieldAccountInfo1,
@@ -152,26 +138,15 @@ public class HttpShieldTrc20Token004 extends ZenTrc20Base {
         shieldAccountInfo1.getString("payment_address"), getRcm(httpnode));
     shieldedReceives = getHttpShieldedReceivesJsonArray(shieldedReceives, account2Receive1V2Amount,
         shieldAccountInfo2.getString("payment_address"), getRcm(httpnode));
-    response = createShieldContractParametersWithoutAskForTransfer(httpnode, shieldAccountInfo1,
-        shieldSpends, shieldedReceives);
-    JSONObject shieldedTrc20Parameters = HttpMethed.parseResponseContent(response);
-    HttpMethed.printJsonContent(shieldedTrc20Parameters);
-    JSONObject spendAuthSig1 = createSpendAuthSig(httpnode, shieldAccountInfo1,
-        shieldedTrc20Parameters.getString("message_hash"), account1IvkNoteTxs.getJSONObject(1)
-            .getJSONObject("note").getString("rcm"));
-    HttpMethed.printJsonContent(spendAuthSig1);
-    JSONArray spendAuthSigArray = new JSONArray();
-    spendAuthSigArray.add(spendAuthSig1);
-    //spendAuthSigArray.add(spendAuthSig2);
-
-    response = getTriggerInputForShieldedTrc20Contract(httpnode, shieldedTrc20Parameters,
-        spendAuthSigArray);
+    response = createShieldContractParametersForTransfer(httpnode, shieldAccountInfo1, shieldSpends,
+        shieldedReceives);
     responseContent = HttpMethed.parseResponseContent(response);
     HttpMethed.printJsonContent(responseContent);
+    Assert.assertTrue(responseContent.containsKey("trigger_contract_input"));
 
     txid = HttpMethed.triggerContractGetTxidWithVisibleTrue(httpnode,anotherHttpnode,
         zenTrc20TokenOwnerAddressString, shieldAddress, transfer, responseContent
-            .getString("value"), maxFeeLimit, 0L, 0, 0L,
+            .getString("trigger_contract_input"), maxFeeLimit, 0L, 0, 0L,
         zenTrc20TokenOwnerKey);
     HttpMethed.waitToProduceOneBlock(httpnode);
     HttpMethed.waitToProduceOneBlock(httpnode);
@@ -190,8 +165,8 @@ public class HttpShieldTrc20Token004 extends ZenTrc20Base {
   }
 
 
-  @Test(enabled = true, description = "Transfer type with 2V2 without ask by http")
-  public void test03TransferTypeWith2V2WithoutAskByHttp() {
+  @Test(enabled = true, description = "Transfer type with 2V2 by http")
+  public void test03TransferTypeWith2V2ByHttp() {
     account2IvkNoteTxs = scanShieldTrc20NoteByIvk(httpnode, shieldAccountInfo2);
 
     Assert.assertFalse(isShieldedTrc20ContractNoteSpent(httpnode, shieldAccountInfo2,
@@ -208,32 +183,16 @@ public class HttpShieldTrc20Token004 extends ZenTrc20Base {
         shieldAccountInfo1.getString("payment_address"), getRcm(httpnode));
     shieldedReceives = getHttpShieldedReceivesJsonArray(shieldedReceives, account2Receive2V2Amount,
         shieldAccountInfo2.getString("payment_address"), getRcm(httpnode));
-    response = createShieldContractParametersWithoutAskForTransfer(httpnode, shieldAccountInfo2,
-        shieldSpends, shieldedReceives);
-    JSONObject shieldedTrc20Parameters = HttpMethed.parseResponseContent(response);
-
-    JSONObject spendAuthSig1 = createSpendAuthSig(httpnode, shieldAccountInfo2,
-        shieldedTrc20Parameters.getString("message_hash"), account2IvkNoteTxs.getJSONObject(0)
-            .getJSONObject("note").getString("rcm"));
-    HttpMethed.printJsonContent(spendAuthSig1);
-
-    JSONObject spendAuthSig2 = createSpendAuthSig(httpnode, shieldAccountInfo2,
-        shieldedTrc20Parameters.getString("message_hash"), account2IvkNoteTxs.getJSONObject(1)
-            .getJSONObject("note").getString("rcm"));
-    JSONArray spendAuthSigArray = new JSONArray();
-    spendAuthSigArray.add(spendAuthSig1);
-    spendAuthSigArray.add(spendAuthSig2);
-
-    response = getTriggerInputForShieldedTrc20Contract(httpnode, shieldedTrc20Parameters,
-        spendAuthSigArray);
+    response = createShieldContractParametersForTransfer(httpnode, shieldAccountInfo2, shieldSpends,
+        shieldedReceives);
     responseContent = HttpMethed.parseResponseContent(response);
     HttpMethed.printJsonContent(responseContent);
+    Assert.assertTrue(responseContent.containsKey("trigger_contract_input"));
 
     txid = HttpMethed.triggerContractGetTxidWithVisibleTrue(httpnode,anotherHttpnode,
         zenTrc20TokenOwnerAddressString, shieldAddress, transfer, responseContent
-            .getString("value"), maxFeeLimit, 0L, 0, 0L,
+            .getString("trigger_contract_input"), maxFeeLimit, 0L, 0, 0L,
         zenTrc20TokenOwnerKey);
-
     HttpMethed.waitToProduceOneBlock(httpnode);
     HttpMethed.waitToProduceOneBlock(httpnode);
     response = HttpMethed.getTransactionInfoById(httpnode, txid, true);
@@ -242,7 +201,8 @@ public class HttpShieldTrc20Token004 extends ZenTrc20Base {
     Assert.assertTrue(responseContent.getJSONObject("receipt")
         .getLong("energy_usage_total") > 300000L);
     Assert.assertEquals(responseContent.getString("contract_address"), shieldAddress);
-    Assert.assertEquals(responseContent.getJSONObject("receipt").getString("result"), "SUCCESS");
+    Assert.assertEquals(responseContent.getJSONObject("receipt").getString("result"),
+        "SUCCESS");
 
     account2OvkNoteTxs = scanShieldTrc20NoteByOvk(httpnode, shieldAccountInfo2);
     logger.info(account1OvkNoteTxs.toJSONString());
@@ -255,8 +215,8 @@ public class HttpShieldTrc20Token004 extends ZenTrc20Base {
 
   }
 
-  @Test(enabled = true, description = "Transfer type with 2V1 without ask by http")
-  public void test04TransferTypeWith2V1WithoutAskByHttp() {
+  @Test(enabled = true, description = "Transfer type with 2V1 by http")
+  public void test04TransferTypeWith2V1ByHttp() {
     account1IvkNoteTxs = scanShieldTrc20NoteByIvk(httpnode, shieldAccountInfo1);
 
     Assert.assertFalse(isShieldedTrc20ContractNoteSpent(httpnode, shieldAccountInfo1,
@@ -274,33 +234,17 @@ public class HttpShieldTrc20Token004 extends ZenTrc20Base {
     shieldedReceives.clear();
     shieldedReceives = getHttpShieldedReceivesJsonArray(shieldedReceives, account1Receive2V1Amount,
         shieldAccountInfo1.getString("payment_address"), getRcm(httpnode));
-    response = createShieldContractParametersWithoutAskForTransfer(httpnode, shieldAccountInfo1,
-        shieldSpends, shieldedReceives);
-    JSONObject shieldedTrc20Parameters = HttpMethed.parseResponseContent(response);
-
-    JSONObject spendAuthSig1 = createSpendAuthSig(httpnode, shieldAccountInfo1,
-        shieldedTrc20Parameters.getString("message_hash"), account1IvkNoteTxs.getJSONObject(2)
-            .getJSONObject("note").getString("rcm"));
-    HttpMethed.printJsonContent(spendAuthSig1);
-
-    JSONObject spendAuthSig2 = createSpendAuthSig(httpnode, shieldAccountInfo1,
-        shieldedTrc20Parameters.getString("message_hash"), account1IvkNoteTxs.getJSONObject(3)
-            .getJSONObject("note").getString("rcm"));
-    HttpMethed.printJsonContent(spendAuthSig2);
-
-    JSONArray spendAuthSigArray = new JSONArray();
-    spendAuthSigArray.add(spendAuthSig1);
-    spendAuthSigArray.add(spendAuthSig2);
-
-    response = getTriggerInputForShieldedTrc20Contract(httpnode, shieldedTrc20Parameters,
-        spendAuthSigArray);
+    response = createShieldContractParametersForTransfer(httpnode, shieldAccountInfo1, shieldSpends,
+        shieldedReceives);
     responseContent = HttpMethed.parseResponseContent(response);
     HttpMethed.printJsonContent(responseContent);
+    Assert.assertTrue(responseContent.containsKey("trigger_contract_input"));
 
     txid = HttpMethed.triggerContractGetTxidWithVisibleTrue(httpnode,anotherHttpnode,
         zenTrc20TokenOwnerAddressString, shieldAddress, transfer, responseContent
-            .getString("value"), maxFeeLimit, 0L, 0, 0L,
+            .getString("trigger_contract_input"), maxFeeLimit, 0L, 0, 0L,
         zenTrc20TokenOwnerKey);
+
     HttpMethed.waitToProduceOneBlock(httpnode);
     HttpMethed.waitToProduceOneBlock(httpnode);
     response = HttpMethed.getTransactionInfoById(httpnode, txid, true);
@@ -309,7 +253,8 @@ public class HttpShieldTrc20Token004 extends ZenTrc20Base {
     Assert.assertTrue(responseContent.getJSONObject("receipt")
         .getLong("energy_usage_total") > 300000L);
     Assert.assertEquals(responseContent.getString("contract_address"), shieldAddress);
-    Assert.assertEquals(responseContent.getJSONObject("receipt").getString("result"), "SUCCESS");
+    Assert.assertEquals(responseContent.getJSONObject("receipt").getString("result"),
+        "SUCCESS");
 
     account1OvkNoteTxs = scanShieldTrc20NoteByOvk(httpnode, shieldAccountInfo1);
     logger.info(account1OvkNoteTxs.toJSONString());
@@ -318,52 +263,20 @@ public class HttpShieldTrc20Token004 extends ZenTrc20Base {
     account1IvkNoteTxs = scanShieldTrc20NoteByIvk(httpnode, shieldAccountInfo1);
     Assert.assertTrue(isShieldedTrc20ContractNoteSpent(httpnode, shieldAccountInfo1,
         account1IvkNoteTxs.getJSONObject(2)));
-    Assert.assertTrue(isShieldedTrc20ContractNoteSpentOnPbft(httpPbftNode, shieldAccountInfo1,
-            account1IvkNoteTxs.getJSONObject(2)));
     Assert.assertTrue(isShieldedTrc20ContractNoteSpent(httpnode, shieldAccountInfo1,
         account1IvkNoteTxs.getJSONObject(3)));
-    Assert.assertTrue(isShieldedTrc20ContractNoteSpentOnPbft(httpPbftNode, shieldAccountInfo1,
-            account1IvkNoteTxs.getJSONObject(3)));
 
   }
 
 
-  @Test(enabled = true, description = "Scan note by ivk and ovk on solidity and pbft by http")
-  public void test05ScanNoteByIvkAndOvkOnSOlidityAndPbftByHttp() {
+  @Test(enabled = true, description = "Query is shielded trc20 contract note spent on "
+      + "solidity by http")
+  public void test05QueryIsShieldedTrc20ContractNoteSpentByHttp() {
     HttpMethed.waitToProduceOneBlockFromSolidity(httpnode, httpSolidityNode);
-
-    account1IvkNoteTxs = scanShieldTrc20NoteByIvk(httpnode, shieldAccountInfo1);
-    JSONArray account1IvkNoteTxsOnSolidity = scanShieldTrc20NoteByIvkOnSolidity(httpSolidityNode,
-        shieldAccountInfo1);
-    Assert.assertEquals(account1IvkNoteTxs, account1IvkNoteTxsOnSolidity);
-    JSONArray account1IvkNoteTxsOnPbft = scanShieldTrc20NoteByIvkOnPbft(httpPbftNode,
-            shieldAccountInfo1);
-    Assert.assertEquals(account1IvkNoteTxs, account1IvkNoteTxsOnPbft);
-
-    account1OvkNoteTxs = scanShieldTrc20NoteByOvk(httpnode, shieldAccountInfo1);
-    JSONArray account1OvkNoteTxsOnSolidity = scanShieldTrc20NoteByOvkOnSolidity(httpSolidityNode,
-        shieldAccountInfo1);
-    Assert.assertEquals(account1OvkNoteTxs, account1OvkNoteTxsOnSolidity);
-    JSONArray account1OvkNoteTxsOnPbft = scanShieldTrc20NoteByOvkOnPbft(httpPbftNode,
-            shieldAccountInfo1);
-    Assert.assertEquals(account1OvkNoteTxs, account1OvkNoteTxsOnPbft);
-
-    account2IvkNoteTxs = scanShieldTrc20NoteByIvk(httpnode, shieldAccountInfo2);
-    JSONArray account2IvkNoteTxsOnSolidity = scanShieldTrc20NoteByIvkOnSolidity(httpSolidityNode,
-        shieldAccountInfo2);
-    Assert.assertEquals(account2IvkNoteTxs, account2IvkNoteTxsOnSolidity);
-    JSONArray account2IvkNoteTxsOnPbft = scanShieldTrc20NoteByIvkOnPbft(httpPbftNode,
-            shieldAccountInfo2);
-    Assert.assertEquals(account2IvkNoteTxs, account2IvkNoteTxsOnPbft);
-
-    account2OvkNoteTxs = scanShieldTrc20NoteByOvk(httpnode, shieldAccountInfo2);
-    JSONArray account2OvkNoteTxsOnSolidity = scanShieldTrc20NoteByOvkOnSolidity(httpSolidityNode,
-        shieldAccountInfo2);
-    Assert.assertEquals(account2OvkNoteTxs, account2OvkNoteTxsOnSolidity);
-    JSONArray account2OvkNoteTxsOnPbft = scanShieldTrc20NoteByOvkOnPbft(httpPbftNode,
-            shieldAccountInfo2);
-    Assert.assertEquals(account2OvkNoteTxs, account2OvkNoteTxsOnPbft);
-
+    Assert.assertTrue(isShieldedTrc20ContractNoteSpentOnSolidity(httpSolidityNode,
+        shieldAccountInfo1, account1IvkNoteTxs.getJSONObject(2)));
+    Assert.assertTrue(isShieldedTrc20ContractNoteSpentOnSolidity(httpSolidityNode,
+        shieldAccountInfo1, account1IvkNoteTxs.getJSONObject(3)));
   }
